pyproject.toml
[project]
name = "bdagcp"
version = "0.1.0"
description = "IAGENERATIVA GCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "functions-framework>=3.10.0",
    "google-cloud-aiplatform>=1.133.0",
    "gradio>=6.2.0",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.2",
    "langchain-google-vertexai>=3.2.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "logging>=0.4.9.6",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
    "teradatasql>=20.0.0.50",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=7.1.0",
    "pyzmq>=27.1.0",
]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where =["src"]
include = ["multiagent/MultiAgenteChatbot*","util*","conf*","agent*","agent/AgenteDeBaseDeDatos*"]
